# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List volunteer opportunities
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: accessibility
          in: query
          schema:
            type: string
        - name: effort
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"

    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create a volunteer opportunity (organizations only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityBody"
      responses:
        "201":
          description: Opportunity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get a single opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opportunity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update an opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityBody"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete an opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /opportunities/{id}/apply:
    post:
      operationId: applyToOpportunity
      tags: [applications]
      summary: Apply to a volunteer opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        organizationId:
          type: integer
        organizationName:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        accessibilityFeatures:
          type: array
          items:
            type: string
        effortLevel:
          type: string
          enum: [low, medium, high]
        location:
          type: string
        state:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        spotsAvailable:
          type: ["integer", "null"]
        isActive:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - organizationId
        - organizationName
        - category
        - skills
        - interests
        - accessibilityFeatures
        - effortLevel
        - location
        - isActive
        - createdAt
    OpportunityDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        organizationId:
          type: integer
        organizationName:
          type: string
        organizationDescription:
          type: ["string", "null"]
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        accessibilityFeatures:
          type: array
          items:
            type: string
        effortLevel:
          type: string
        location:
          type: string
        state:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        spotsAvailable:
          type: ["integer", "null"]
        applicantsCount:
          type: integer
        isActive:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - organizationId
        - organizationName
        - category
        - skills
        - interests
        - accessibilityFeatures
        - effortLevel
        - location
        - isActive
        - applicantsCount
        - createdAt
    CreateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        accessibilityFeatures:
          type: array
          items:
            type: string
        effortLevel:
          type: string
          enum: [low, medium, high]
        location:
          type: string
        state:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        spotsAvailable:
          type: integer
      required:
        - title
        - description
        - category
        - effortLevel
        - location
    UpdateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        accessibilityFeatures:
          type: array
          items:
            type: string
        effortLevel:
          type: string
        location:
          type: string
        state:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        spotsAvailable:
          type: integer
        isActive:
          type: boolean
    ApplyBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Application:
      type: object
      properties:
        id:
          type: integer
        opportunityId:
          type: integer
        opportunityTitle:
          type: string
        userId:
          type: integer
        userName:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
      required:
        - id
        - opportunityId
        - opportunityTitle
        - userId
        - userName
        - message
        - status
        - createdAt
...