# lib/api-zod/src/generated/types/applyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
 * OpenAPI spec version: 0.1.0
 */

export interface ApplyBody {
  message: string;
}