# lib/api-zod/src/generated/types/opportunityDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
 * OpenAPI spec version: 0.1.0
 */

export interface OpportunityDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  organizationId: number;
  organizationName: string;
  /** @nullable */
  organizationDescription?: string | null;
  category: string;
  skills: string[];
  interests: string[];
  accessibilityFeatures: string[];
  effortLevel: string;
  location: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  spotsAvailable?: number | null;
  applicantsCount: number;
  isActive: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}